firmware: Disable interrupts when accessing millisecond counter
The millisecond counter is a software variable but it's a uint16_t.  This
means the access will not be atomic on an 8-bit system.  An interrupt
could happen in between reading the two bytes and change the value.

This patch disables interrupts before a read/reset of the counter, if they
were enabled.  Afterwards, it reenables the interrupt if it was previously
enabled.