firmware: Fix USB library to account for timer overflow
The 16-bit millisecond counter overflows about once a minute (2^16 ms).
The USB library should compute intervals between two values read from the
counter, and compare that to a threshold.  This will properly handle the
wraparound.

Previously, the code computed the expected threshold time (initial +
delay).  This is not robust enough to deal with overflows.